% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assoc_GWAScatalog.R
\name{get_assoc_GWAScatalog}
\alias{get_assoc_GWAScatalog}
\title{Obtain all available associations from a study from GWAScatalog, with GRCh37/hg19 position data}
\usage{
get_assoc_GWAScatalog(study)
}
\arguments{
\item{study}{A GWAScatalog study code}
}
\value{
A tibble
}
\description{
Obtain all available associations from a study from GWAScatalog, with GRCh37/hg19 position data
}
\examples{
get_assoc_GWAScatalog("GCST006259")

}
