% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sumstats_GWAScatalog.R
\name{get_sumstats_GWAScatalog}
\alias{get_sumstats_GWAScatalog}
\title{Obtain summary statistics from a study from GWAScatalog with GRCh37/hg19 position data}
\usage{
get_sumstats_GWAScatalog(study)
}
\arguments{
\item{study}{A GWAScatalog study code}
}
\value{
A tibble
}
\description{
Obtain summary statistics from a study from GWAScatalog with GRCh37/hg19 position data
}
\examples{
get_sumstats_GWAScatalog("GCST006259")

}
