% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_test_p.R
\name{wald_test_p}
\alias{wald_test_p}
\title{Simple calculation of p-value from a Wald test}
\usage{
wald_test_p(beta, se)
}
\arguments{
\item{beta}{Regression coefficient}

\item{se}{Standard error of the model}
}
\value{
P-value (as numeric vector)
}
\description{
Simple calculation of p-value from a Wald test
}
\examples{
data <- lm(mpg ~ cyl, data = mtcars)
beta <- coef(summary(data))[, "Estimate"]["cyl"]
se <- coef(summary(data))[, "Std. Error"]["cyl"]

wald_test_p(beta, se)
}
